requirements.txt
bcrypt==4.2.1
fastapi
httpx==0.28.1
jwt==1.3.1
python-jose==3.3.0
python-multipart==0.0.20
passlib==1.7.4
SQLAlchemy==2.0.37
uvicorn==0.34.0
psycopg2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-multipart==0.0.20
